+++ ptai-jenkins-plugin/src/main/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/plugin/jenkins/actions/AstJobTableResults.java
@@ -31,7 +31,9 @@ public class AstJobTableResults implements Action {
 
     @Override
     public String getIconFileName() {
-        return "plugin/" + Jenkins.get().getPluginManager().getPlugin("ptai-jenkins-plugin").getShortName() + "/24x24.png";
+        // TODO: Implement project actions and uncomment this
+        // return "plugin/" + Jenkins.get().getPluginManager().getPlugin("ptai-jenkins-plugin").getShortName() + "/24x24.png";
+        return null;
     }
 
     @Override